pyproject.toml
[project]
name = "adkx"
description = "Extensions and utilities for Google ADK"
authors = [
    {name = "Wei Sun (Jack)", email = "jacksunwei@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = ["adk", "agent", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "google-adk>=1.18.0",
]
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/Jacksunwei/adkx-python"
repository = "https://github.com/Jacksunwei/adkx-python"
documentation = "https://github.com/Jacksunwei/adkx-python#readme"
changelog = "https://github.com/Jacksunwei/adkx-python/releases"

[project.optional-dependencies]
dev = [
    "flit>=3.10.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "pyink>=24.10.0",
    "pylint>=2.6.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "adkx"

[tool.pyink]
# Format py files following Google style-guide
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
pyink-annotation-pragmas = [
  "noqa",
  "pylint:",
  "type: ignore",
  "pytype:",
  "mypy:",
  "pyright:",
  "pyre-",
]
extend-exclude = '''/(\.venv|build|dist)/'''

[tool.isort]
profile = "google"
single_line_exclusions = []
line_length = 200                  # Prevent line wrap flickering.
known_third_party = ["google.adk"]
